% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_missing_capl_variables}
\alias{add_missing_capl_variables}
\title{Add mandatory CAPL variables to a data frame of raw data if they are missing.}
\usage{
add_missing_capl_variables(raw_data = NULL)
}
\arguments{
\item{raw_data}{a data frame or tibble of raw CAPL data.}
}
\value{
returns a merged data frame of raw data and missing mandatory CAPL variables (values are set to NA).
}
\description{
Add mandatory CAPL variables to a data frame of raw data if they are missing.
}
\examples{
raw_data <- add_missing_capl_variables(raw_data)

}
