% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{capitalize_character}
\alias{capitalize_character}
\title{Capitalize a character vector.}
\usage{
capitalize_character(x = NA)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Returns a character vector (if valid) or NA (if not valid).
}
\description{
This function capitalizes a character vector.
}
\details{
Other \code{capl} functions called by this function include: \code{\link[=validate_character]{validate_character()}}.
}
\examples{
capitalize_character(c("beginning", "progressing", "achieving", "excelling"))

# [1] "Beginning"   "Progressing" "Achieving"   "Excelling"

}
