% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl_demo_data.R
\docType{data}
\name{capl_demo_data}
\alias{capl_demo_data}
\title{CAPL demo raw data.}
\format{
A data frame of 500 observations on 60 variables that are required in order to compute CAPL-2 scores and interpretations:
\describe{
}
}
\source{
\url{https://github.com/barnzilla/capl}
}
\usage{
capl_demo_data
}
\description{
A dataset containing CAPL demo raw data.
}
\keyword{datasets}
