% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl_integer.r
\name{capl_integer}
\alias{capl_integer}
\title{Converts CAPL interpretations from strings to integers.}
\usage{
capl_integer(x)
}
\arguments{
\item{x}{A character vector (beginning, progressing, achieving, excelling).}
}
\value{
An integer vector (1 = beginning, 2 = progressing, 3 = achieving, 4 = excelling).
}
\description{
This function converts CAPL interpretations from strings to integers.
}
\examples{
capl_interpretations <- c("beginning", "progressing", "achieving", "excelling")
capl_integer(capl_interpretations)
}
