% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl_iqr.r
\name{capl_iqr}
\alias{capl_iqr}
\title{Evaluates CAPL data against the IQR rule for outliers.}
\usage{
capl_iqr(x, age, sex, min_values = 25)
}
\arguments{
\item{x}{A numeric vector.}

\item{age}{An integer vector.}

\item{sex}{A vector containing numeric or character values.}

\item{min_values}{An optional parameter that specifies the minimum number of values for a given age- and sex-specific distribution required before the function will evaluate the values within the x argument related to that distribution against the IQR rule for outliers. The default is 25. If a given age- and sex-specific distribution has less values than the minimum, NAs will be returned for all values related to that distribution.}
}
\value{
An integer vector of dichotomous values (0-1) where 1 identifies a value as an outlier based on the IQR rule for outliers.
}
\description{
This function evaluates CAPL data against the IQR rule for outliers based on age- and sex-specific distributions of the data. According to this rule, a value less than quartile 1 - (1.5 * the interquartile range) or greater than quartile 3 + (1.5 * the interquartile range) is considered an outlier.
}
\examples{
# Use of the function with 500 randomly generated values for 8- to 9-year-old males and females
random_values <- runif(500, 1, 100)
random_values[c(1:2, 499:500)] <- c(rep(-100, 2), rep(200, 2))
age <- c(rep(8, 250), rep(9, 250))
sex <- c(rep("male", 250), rep("female", 250))
capl_iqr(random_values, age, sex)
}
