% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl_pedometer.r
\name{capl_pedometer}
\alias{capl_pedometer}
\title{Evaluates whether pedometer steps and wear time qualify as a valid day of pedometry and provides a CAPL pedometer score.}
\usage{
capl_pedometer(steps, wear_time, edition = 2)
}
\arguments{
\item{steps}{An integer vector of pedometer steps.}

\item{wear_time}{A numeric vector of pedometer wear time in decimal hours.}

\item{edition}{An optional integer vector indicating which CAPL edition to use for calculating pedometer scores (1, 2). Default is 2.}
}
\value{
A dataframe with 2 variables: $valid representing an integer vector of dichotomous values (0-1) where 1 identifies a valid day of pedometry, and $score representing an integer vector of values representing pedometer scores based on the CAPL scoring system.
}
\description{
This function evaluates whether pedometer steps and wear time qualify as a valid day of pedometry and provides a pedometer score based on the CAPL scoring system. The rules for a valid day of pedometry are: 1000 <= steps <= 30000 and wear_time >= 10 hours.
}
\examples{
# Use of the function with 1000 randomly generated step and wear time values
steps <- runif(1000, 1000, 30000)
wear_time <- runif(1000, 9.9, 12)
output <- capl_pedometer(steps, wear_time)

output$valid
output$score
}
