% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl_time.r
\name{capl_time}
\alias{capl_time}
\title{Converts CAPL pedometer time on/time off data to 24-hour format.}
\usage{
capl_time(x)
}
\arguments{
\item{x}{A character vector of time values.}
}
\value{
A character vector of time values in 24-hour format (hh:mm).
}
\description{
This function converts time strings to 24-hour format (hh:mm).
}
\examples{
time_values <- c("4:15 PM", "8:00", "20:45", "7pm", "9:23am")
capl_time(time_values)
}
