% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl_wear_time.r
\name{capl_wear_time}
\alias{capl_wear_time}
\title{Calculates pedometer daily wear time.}
\usage{
capl_wear_time(time_on, time_off, non_wear_time = 0)
}
\arguments{
\item{time_on}{A character vector of time values.}

\item{time_off}{A character vector of time values.}

\item{non_wear_time}{An optional integer vector of time values in minutes. Default is 0.}
}
\value{
A numeric vector of pedometer daily wear time values in decimal hours.
}
\description{
This function calculates pedometer daily wear time values and accounts for non-wear time values.
}
\examples{
# Use of the function with time on, time off and non-wear time values
time_on <- c("07:15", "8:30 am", "10:00 AM", "11:30", "7am")
time_off <- c("6:30 pm", "20:59", "23:15", "10:30 PM", "9:45pm")
non_wear_time <- c(25, 60, 0, 0, 1)
capl_wear_time(time_on, time_off, non_wear_time)

# Other examples
time_on <- c("07:15", "8:30 am", "10:00 AM", "11:30", "7am")
time_off <- c("20:59")
capl_wear_time(time_on, time_off)

capl_wear_time(time_on, time_off, 30)
}
