% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl_z.r
\name{capl_z}
\alias{capl_z}
\title{Converts CAPL data to age- and sex-specific z-scores.}
\usage{
capl_z(x, age, sex)
}
\arguments{
\item{x}{A numeric vector.}

\item{age}{An integer vector.}

\item{sex}{A vector containing numeric or character values.}
}
\value{
A vector of age- and sex-specific z-scores.
}
\description{
This function converts CAPL data to z-scores based on age- and sex-specific distributions of the vector.
}
\examples{
# Use of the function with 500 randomly generated values for 8- to 9-year-old males and females
random_values <- runif(500, 1, 100)
random_values[c(1:2, 499:500)] <- c(rep(-100, 2), rep(200, 2))
age <- c(rep(8, 250), rep(9, 250))
sex <- c(rep("male", 250), rep("female", 250))
capl_z(random_values, age, sex)
}
