% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{export_capl_data}
\alias{export_capl_data}
\title{Export CAPL-2 data to an Excel workbook.}
\usage{
export_capl_data(x = NULL, file_path = NA)
}
\arguments{
\item{x}{A data frame.}

\item{file_path}{A character vector representing the file path to a location on the user's local computer (e.g., "c:/users/user_name/desktop/file.xlsx")
where \code{x} will be saved as an Excel workbook on the user's computer. The file path is not case-sensitive.}
}
\description{
This function exports CAPL-2 data to an Excel workbook on a local computer.
}
\details{
Other \code{capl} functions called by this function include: \code{\link[=validate_character]{validate_character()}}.
}
