% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_competence.R
\name{get_best_camsa_score}
\alias{get_best_camsa_score}
\title{Compute the maximum CAMSA (Canadian Agility and Movement Skill Assessment) skill + time score out of two trials.}
\usage{
get_best_camsa_score(camsa_score1 = NA, camsa_score2 = NA)
}
\arguments{
\item{camsa_score1}{a numeric element or vector representing the skill + time score from trial 1 (valid values are between 1 and 28).}

\item{camsa_score2}{a numeric element or vector representing the skill + time score from trial 2 (valid values are between 1 and 28).}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Compute the maximum CAMSA (Canadian Agility and Movement Skill Assessment) skill + time score out of two trials.
}
\examples{
get_best_camsa_score(
  camsa_score1 = c(1, 5, 10, 28, 29), 
  camsa_score2 = c(5, 7, 12, NA, 27)
)
# [1]  5  7 12 NA NA

}
