% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_competence.R
\name{get_camsa_time_score}
\alias{get_camsa_time_score}
\title{Compute the CAMSA (Canadian Agility and Movement Skill Assessment) time score based on the time taken to complete a trial.}
\usage{
get_camsa_time_score(camsa_time = NA)
}
\arguments{
\item{camsa_time}{a numeric element or vector representing the time (in seconds) taken to complete a CAMSA trial (valid values are > 0).}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Compute the CAMSA (Canadian Agility and Movement Skill Assessment) time score based on the time taken to complete a trial.
}
\examples{
get_camsa_time_score(c(14, 12, 30, 25, 0))

# [1] 13 14  1  4 NA

}
