% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_behaviour.R
\name{get_db_score}
\alias{get_db_score}
\title{Compute a daily behaviour domain score based on the step and self-reported physical activity scores.}
\usage{
get_db_score(step_score = NA, self_report_pa_score = NA)
}
\arguments{
\item{step_score}{a numeric (integer) element or vector representing the pedometer steps protocol score (valid values are integers between 0 and 25).}

\item{self_report_pa_score}{a numeric (integer) element or vector representing the self-reported physical activity protocol score (valid values are
integers between 0 and 5).}
}
\value{
returns a numeric (integer) element between 0 and 30 (if valid) or NA (if not valid).
}
\description{
Compute a daily behaviour domain score based on the step and self-reported physical activity scores.
}
\examples{
get_db_score(
  step_score = c(20, 6, 13, 5, NA, 4.5),
  self_report_pa_score = c(3, 2, 1, 4, 7, 3)
)

# [1] 23  8 14  9 NA NA

}
