% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivation_and_confidence.R
\name{get_mc_score}
\alias{get_mc_score}
\title{Compute a motivation and confidence domain score based on the step and self-reported physical activity scores.}
\usage{
get_mc_score(
  predilection_score = NA,
  adequacy_score = NA,
  intrinsic_motivation_score = NA,
  pa_competence_score = NA
)
}
\arguments{
\item{predilection_score}{a numeric element or vector (valid values are between 1.8 and 7.5).}

\item{adequacy_score}{a numeric element or vector (valid values are between 1.8 and 7.5).}

\item{intrinsic_motivation_score}{a numeric element or vector (valid values are between 1.5 and 7.5).}

\item{pa_competence_score}{a numeric element or vector (valid values are between 1.5 and 7.5).}
}
\value{
returns a numeric element between 0 and 30 (if valid) or NA (if not valid).
}
\description{
Compute a motivation and confidence domain score based on the predilection, adequacy, intrinsic motivation and physical activity
competence scores. If one of the scores is missing or invalid, a weighted domain score will be computed from the other three scores.
}
\details{
This function calls \code{\link[=validate_number]{validate_number()}} and \code{\link[=validate_protocol_score]{validate_protocol_score()}} to ensure that the computed score is valid.
}
\examples{
get_mc_score(
  predilection_score = c(7, 7.5, 5, 8, 4),
  adequacy_score = c(NA, 5, 3, 1, 4),
  intrinsic_motivation_score = c(5, 7.5, 4, 2, 3.5),
  pa_competence_score = c(6, 3, 6, 7, 2)
)

# [1] 24.0 23.0 18.0   NA 13.5

}
