% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_missing_capl_variables}
\alias{get_missing_capl_variables}
\title{Add required CAPL variables to a data frame of raw data if they are missing.}
\usage{
get_missing_capl_variables(raw_data = NULL)
}
\arguments{
\item{raw_data}{a data frame or tibble of raw CAPL data.}
}
\value{
returns a merged data frame of raw data and missing required CAPL variables (values are set to NA).
}
\description{
Add required CAPL variables to a data frame of raw data if they are missing.
}
\examples{
raw_data <- get_missing_capl_variables(raw_data)

}
