% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivation_and_confidence.R
\name{get_pa_competence_score}
\alias{get_pa_competence_score}
\title{Compute a score for responses to BREQ items 4-6.}
\usage{
get_pa_competence_score(
  feelings_about_pa1 = NA,
  feelings_about_pa2 = NA,
  feelings_about_pa3 = NA
)
}
\arguments{
\item{feelings_about_pa1}{a numeric (integer) element or vector representing a response to BREQ item 4 (valid values are integers between 1 and 5).}

\item{feelings_about_pa2}{a numeric (integer) element or vector representing a response to BREQ item 5 (valid values are integers between 1 and 5).}

\item{feelings_about_pa3}{a numeric (integer) element or vector representing a response to BREQ item 6 (valid values are integers between 1 and 5).}
}
\value{
returns a numeric element between 1.5 and 7.5 (if valid) or NA (if not valid).
}
\description{
This function computes an intrinsic motivation score for responses to items 4-6 of the the Behavioral Regulation in Exercise Questionnaire (BREQ) as they
appear in the CAPL-2 questionnaire. This score is used to compute the motivation and confidence domain score.
}
\details{
This function calls \code{\link[=validate_questionnaire_item]{validate_questionnaire_item()}} to validate the arguments before computing the intrinsic motivation score.

Valid elements (integers between 1 and 5) represent the following responses:
\itemize{
\item 1 = Not true for me
\item 2 = Not really true for me
\item 3 = Sometimes true for me
\item 4 = Often true for me
\item 5 = Very true for me
}
}
\examples{
get_pa_competence_score(
  feelings_about_pa1 = c(4, 3, 6, 5, "2"),
  feelings_about_pa2 = c(1:5),
  feelings_about_pa3 = c(1, 5, 4, 3, 3)
)

# [1]  3  5 NA  6  5

}
