% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_competence.R
\name{get_pc_score}
\alias{get_pc_score}
\title{Compute a physical competence domain score based on the PACER, plank and CAMSA scores. If one protocol score is missing or invalid, a weighted domain
score will be computed from the other two protocol scores.}
\usage{
get_pc_score(pacer_score = NA, plank_score = NA, camsa_overall_score = NA)
}
\arguments{
\item{pacer_score}{a numeric element or vector representing the PACER protocol score (valid values are between 0 and 10).}

\item{plank_score}{a numeric element or vector representing the PACER protocol score (valid values are between 0 and 10).}

\item{camsa_overall_score}{a numeric element or vector representing the best CAMSA skill + skill score divided by 2.8 (valid values are between 0 and 10).}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Compute a physical competence domain score based on the PACER, plank and CAMSA scores. If one protocol score is missing or invalid, a weighted domain
score will be computed from the other two protocol scores.
}
\examples{
get_pc_score(
  pacer_score = c(1, 5, 8, 10, NA),
  plank_score = c(4, 5, 5, 6, 9),
  camsa_overall_score = c(-1, 0, 6, 4, 3)
)
# [1]  7.5 10.0 19.0 20.0 18.0

}
