% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_competence.R
\name{get_plank_score}
\alias{get_plank_score}
\title{Compute a plank score based on the number of seconds a plank was held.}
\usage{
get_plank_score(plank_time = NA)
}
\arguments{
\item{plank_time}{a numeric element or vector representing time (seconds).}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Compute a plank score based on the number of seconds a plank was held.
}
\examples{
get_plank_score(c(120.5, 75.6, 40, 10.99, 90))
# [1] 10  6  3  0  8

}
