% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_behaviour.R
\name{get_step_average}
\alias{get_step_average}
\title{Compute average daily steps taken.}
\usage{
get_step_average(raw_data = NULL)
}
\arguments{
\item{raw_data}{A data frame that includes seven days of pedometer steps (steps1...steps7) and their corresponding on (time_on1...time_on7)
and off (time_off1...time_off7) times.}
}
\value{
Returns a data frame with nine columns: \code{steps1} (validated), \code{steps2} (validated), \code{steps3} (validated), \code{steps4} (validated), \code{steps5} (validated), \code{steps6} (validated), \code{steps6} (validated), \code{valid_days} and \code{step_average}. The steps are validated with the \code{\link[=validate_steps]{validate_steps()}} function.
}
\description{
This function computes the daily arithmetic mean of a week of steps taken as measured by a pedometer.
}
\details{
There must be at least three valid days for an arithmetic mean to be computed. If only three valid days, one of the step values from a valid
day will be randomly sampled and used for the fourth valid day before computing the mean.

Other \code{capl} functions called by this function include: \code{\link[=validate_steps]{validate_steps()}} and \code{\link[=get_pedometer_wear_time]{get_pedometer_wear_time()}}.
}
\examples{
capl_demo_data <- get_capl_demo_data(10)

get_step_average(capl_demo_data)$step_average

# [1] 18365 12655 15493 12966 11396 13954 18456 13589 17543 11276

}
