% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_behaviour.R
\name{get_valid_days}
\alias{get_valid_days}
\title{Sum the number of valid days that a pedometer was worn.}
\usage{
get_valid_days(x = NULL)
}
\arguments{
\item{x}{a data frame or tibble with seven columns where each column represents a day of the week when a pedometer was worn and contains "yes" or
"no" values. See \code{\link[=validate_steps]{validate_steps()}}.}
}
\value{
returns a numeric element (if valid) or 0 (if not valid).
}
\description{
Sum the number of valid days that a pedometer was worn.
}
\examples{
x <- data.frame(
  valid_steps1 = c("yes", "no", "yes"),
  valid_steps2 = c("no", "no", "yes"),
  valid_steps3 = c("yes", "yes", "yes"),
  valid_steps4 = c(NA, "no", "yes"),
  valid_steps5 = c("yes", "", "yes"),
  valid_steps6 = c("no", "yes", "yes"),
  valid_steps7 = c("yes", "yes", "no")
)

get_valid_days(x)

# [1] 4 3 6

}
