% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivation_and_confidence.R
\name{validate_breq}
\alias{validate_breq}
\title{Check whether a response to a given BREQ item is valid.}
\usage{
validate_breq(x)
}
\arguments{
\item{x}{a numeric (integer) element or vector representing the response to a BREQ question (valid values are between 1 and 5).}
}
\value{
returns a numeric (integer) element (if valid) or NA (if not valid).
}
\description{
This function checks whether an element for a given item on the Behavioral Regulation in Exercise Questionnaire (BREQ) is valid. A valid element is an
integer between 1 and 5 (see Details for more information). This function calls \code{\link[=validate_integer]{validate_integer()}} and is called by \code{\link[=get_intrinsic_motivation_score]{get_intrinsic_motivation_score()}}
and \code{\link[=get_pa_competence_score]{get_pa_competence_score()}}.
}
\details{
Valid elements (integers between 1 and 5) represent the following responses:
\itemize{
\item 1 = Not true for me
\item 2 = Not really true for me
\item 3 = Sometimes true for me
\item 4 = Often true for me
\item 5 = Very true for me
}
}
\examples{
validate_breq(c(1:5, 5.5, NA, "", "3"))

# [[1]  1  2  3  4  5 NA NA NA  3

}
