% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivation_and_confidence.R
\name{validate_csappa}
\alias{validate_csappa}
\title{Check whether a response to a given CSAPPA item is valid.}
\usage{
validate_csappa(x)
}
\arguments{
\item{x}{a numeric (integer) element or vector representing the response to a CSAPPA item (valid values are between 1 and 4).}
}
\value{
returns a numeric (integer) element (if valid) or NA (if not valid).
}
\description{
This function checks whether an element for a given item on the Children's Self-Perceptions of Adequacy in and Predilection for Physical Activity
(CSAPPA; #' Hay, 1992) questionnaire is valid. A valid element is an integer between 1 and 4 (see Details for more information). This function calls
\code{\link[=validate_integer]{validate_integer()}} and is called by \code{\link[=get_adequacy_score]{get_adequacy_score()}} and \code{\link[=get_predilection_score]{get_predilection_score()}}.
}
\details{
Valid elements (integers between 1 and 4) represent the following responses:
\itemize{
\item 1 = REALLY TRUE for me for "some kids" statements
\item 2 = SORT OF TRUE for me for "some kids" statements
\item 3 = REALLY TRUE for me for "other kids" statements
\item 4 = SORT OF TRUE for me for "other kids" statements
}
}
\examples{
validate_csappa(c(1:4, 5.5, NA, "", "3"))

# [1]  1  2  3  4 NA NA NA  3

}
