% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_gender}
\alias{validate_gender}
\title{Check whether a vector can be classified as "girl" or "boy".}
\usage{
validate_gender(x)
}
\arguments{
\item{x}{A vector (see Examples for valid values).}
}
\value{
Returns a character vector with values of "girl" or "boy" (if valid) or NA (if not valid).
}
\description{
This function checks whether a vector can be classified as "girl" or "boy".
}
\examples{
validate_gender(c("Girl", "GIRL", "g", "G", "Female", "f", "F", "", NA, 1))

# [1] "girl" "girl" "girl" "girl" "girl" "girl" "girl" NA     NA     "girl"

validate_gender(c("Boy", "BOY", "b", "B", "Male", "m", "M", "", NA, 0))

# [1] "boy" "boy" "boy" "boy" "boy" "boy" "boy" NA    NA    "boy"

}
