% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_integer}
\alias{validate_integer}
\title{Check whether an element is an integer.}
\usage{
validate_integer(x)
}
\arguments{
\item{x}{an element or vector.}
}
\value{
returns a numeric (integer) element (if valid) or NA (if not valid).
}
\description{
Check whether an element is an integer.
}
\examples{
validate_integer(c(2, 6, 3.3, "", NA, "6", "hello, world"))

# [1]  2  6 NA NA NA  6 NA

}
