% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_number}
\alias{validate_number}
\title{Check whether an element is numeric.}
\usage{
validate_number(x)
}
\arguments{
\item{x}{an element or vector.}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Check whether an element is numeric.
}
\examples{
validate_number(c(1:5, "5", "", NA, "hello, world!"))
# [1]  1  2  3  4  5  5 NA NA NA

}
