% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_pc_protocol_score}
\alias{validate_pc_protocol_score}
\title{Check whether a physical competence score is numeric and between 0 and 10}
\usage{
validate_pc_protocol_score(x)
}
\arguments{
\item{x}{an element or vector representing a physical competence protocol score.}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Check whether a physical competence score is numeric and between 0 and 10
}
\examples{
validate_pc_protocol_score(c(1, 5, 10, 11, -1, NA, "6"))
# [1]  1  5 10 NA NA NA  6

}
