% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_protocol_score}
\alias{validate_protocol_score}
\title{Check whether a CAPL protocol score is numeric and within a valid range.}
\usage{
validate_protocol_score(x = NA, protocol = NA)
}
\arguments{
\item{x}{an element or vector representing a CAPL protocol score.}

\item{protocol}{a character element representing protocols within one of the four CAPL domains (valid values currently include "pc"; valid values are
not case-sensitive).}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Check whether a CAPL protocol score is numeric and within a valid range.
}
\examples{
validate_protocol_score(
  x = c(1, 5, 10, 11, -1, NA, "6"),
  protocol = "pc"
)
# [1]  1  5 10 NA NA NA  6

}
