% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_questionnaire_item}
\alias{validate_questionnaire_item}
\title{Check whether a response to a given questionnaire item is valid.}
\usage{
validate_questionnaire_item(x, lower_bound = NA, upper_bound = NA)
}
\arguments{
\item{x}{a numeric (integer) element or vector representing the response to a questionnaire item (valid values are between the values set by the
lower_bound and upper_bound argumetns).
#param lower_bound a numeric element representing the value below which x is invalid.
#param upper_bound a numeric element representing the value above which x is invalid.}
}
\value{
returns a numeric (integer) element (if valid) or NA (if not valid).
}
\description{
This function checks whether an element for a given questionnaire item is valid. This function calls
\code{\link[=validate_integer]{validate_integer()}} and \code{\link[=validate_number]{validate_number()}} and is called by \code{\link[=get_adequacy_score]{get_adequacy_score()}} and \code{\link[=get_predilection_score]{get_predilection_score()}}.
}
\examples{
validate_questionnaire_item(
  x = c(0:10, NA, "7"),
  lower_bound = 1,
  upper_bound = 7
)

# [1] NA  1  2  3  4  5  6  7 NA NA NA NA  7

}
