% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_behaviour.R
\name{validate_self_report_pa}
\alias{validate_self_report_pa}
\title{Check whether a response to the self-reported physical activity question is valid.}
\usage{
validate_self_report_pa(x)
}
\arguments{
\item{x}{a numeric (integer) element or vector representing the steps taken on a given day (valid values are between 0 and 7).}
}
\value{
returns a numeric (integer) element (if valid) or NA (if not valid).
}
\description{
This function checks whether the response to "During the past week (7 days), on how many days were you physically active for a total of at least 60
minutes per day? (all the time you spent in activities that increased your heart rate and made you breathe hard)?" is valid. A valid response is an integer
between 0 and 7. Data should be processed using this function before calling \code{\link[=get_self_report_pa_score]{get_self_report_pa_score()}}.
}
\examples{
validate_self_report_pa(c(1:8, 5.5, NA, "", "5"))

# [1]  1  2  3  4  5  6  7 NA NA NA NA  5

}
