% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl_demo_data.R
\docType{data}
\name{capl_demo_data}
\alias{capl_demo_data}
\title{CAPL demo raw data.}
\format{
A data frame with 500 rows of data on 60 variables that are required to compute CAPL-2 scores and interpretations:
\describe{
...
}
}
\source{
\url{https://github.com/barnzilla/capl}
}
\usage{
capl_demo_data
}
\description{
A dataset containing CAPL-2 demo raw data.
}
\keyword{datasets}
