% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_24_hour_clock}
\alias{get_24_hour_clock}
\title{Convert a 12-hour clock element to a 24-hour clock element.}
\usage{
get_24_hour_clock(x = NA)
}
\arguments{
\item{x}{a character element or vector in 12-hour clock format.}
}
\value{
returns a 24-hour clock element (if valid) or the original element as a character element (if not valid).
}
\description{
Convert a 12-hour clock element to a 24-hour clock element.
}
\examples{
get_24_hour_clock(c("5:00 am", "7:10PM", "21:37", NA, "", 9))
# [1] "05:00" "19:10" "21:37" NA      NA      "9" 

}
