% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_competence.R
\name{get_camsa_score}
\alias{get_camsa_score}
\title{Compute the CAMSA (Canadian Agility and Movement Skill Assessment) skill + time score.}
\usage{
get_camsa_score(camsa_skill_score = NA, camsa_time_score = NA)
}
\arguments{
\item{camsa_skill_score}{a numeric element or vector representing the CAMSA skill score (valid values are between 0 and 14).}

\item{camsa_time_score}{a numeric element or vector representing the CAMSA time score (valid values are between 1 and 14).}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Compute the CAMSA (Canadian Agility and Movement Skill Assessment) skill + time score.
}
\examples{
get_camsa_score(
  camsa_skill_score = c(0, 5, 10, 14, 15),
  camsa_time_score = c(1, 10, 12, 15, 30)
)
# [1]  1 15 22 NA NA

}
