% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl.R
\name{get_capl}
\alias{get_capl}
\title{The main function in this package that computes CAPL scores and interpretations from raw data.}
\usage{
get_capl(raw_data = NULL)
}
\arguments{
\item{raw_data}{a data frame or tibble of raw CAPL data.}
}
\value{
returns a data frame of CAPL scores and interpretations.
}
\description{
The main function in this package that computes CAPL scores and interpretations from raw data.
}
