% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl.R
\name{get_capl}
\alias{get_capl}
\title{The main (wrapper) function in the CAPL package that computes CAPL scores and interpretations from raw data.}
\usage{
get_capl(raw_data = NULL, sort = "abc")
}
\arguments{
\item{raw_data}{a data frame or tibble of raw CAPL data.}

\item{sort}{a character element representing how the variables in the returned data frame should be sorted (valid values are "abc", "zyx" and "asis"; valid
values are not case-sensitive). This argument is set to "abc" by default.}
}
\value{
returns a merged data frame of raw data and CAPL scores and interpretations.
}
\description{
The main (wrapper) function in the CAPL package that computes CAPL scores and interpretations from raw data.
}
\examples{
get_capl(raw_data)

}
