% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capl.R
\name{get_capl_interpretation}
\alias{get_capl_interpretation}
\title{Compute an age- and gender-specific CAPL interpretation for a given CAPL protocol or domain score.}
\usage{
get_capl_interpretation(age = NA, gender = NA, score = NA, protocol = NA)
}
\arguments{
\item{age}{a numeric element or vector (valid values are between 8 and 12).}

\item{gender}{a character element or vector (valid values currently include "girl", "g", "female", "f", "boy", "b", "male", "m").}

\item{score}{a numeric element or vector. If the protocol argument is set to "pacer" or "steps", the element(s) in this argument must be integers.}

\item{protocol}{a character element representing a CAPL protocol (valid values currently include "pacer", "plank", "camsa", "pc", "steps",
"self_report_pa", "db", "mc"; valid values are not case-sensitive).}
}
\value{
returns a character element with a value of "beginning", "progressing", "achieving" or "excelling" (if valid) or NA (if not valid).
}
\description{
Compute an age- and gender-specific CAPL interpretation for a given CAPL protocol or domain score.
}
\examples{
get_capl_interpretation(
  age = 7:13, 
  gender = c("g", "g", "b", "Boy", "m", "f", "Female"), 
  score = c(50, 25, 100, 5, 150, 23, 78), 
  protocol = "pacer"
)

# [1] NA            "achieving"   "excelling"   "beginning"   "excelling"   "progressing"
# [7] NA

}
