% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_and_understanding.R
\name{get_ku_score}
\alias{get_ku_score}
\title{Compute a knowledge and understanding domain score.}
\usage{
get_ku_score(
  pa_guideline_score = NA,
  cardiorespiratory_fitness_means_score = NA,
  muscular_strength_means_score = NA,
  sports_skill_score = NA,
  fill_in_the_blanks_score = NA
)
}
\arguments{
\item{pa_guideline_score}{a numeric (integer) element or vector (valid values are between 0 and 1).}

\item{cardiorespiratory_fitness_means_score}{a numeric (integer) element or vector (valid values are between 0 and 1).}

\item{muscular_strength_means_score}{a numeric (integer) element or vector (valid values are between 0 and 1).}

\item{sports_skill_score}{a numeric (integer) element or vector (valid values are between 0 and 1).}

\item{fill_in_the_blanks_score}{a numeric (integer) element or vector (valid values are between 0 and 6).}
}
\value{
returns a numeric element between 0 and 10 (if valid) or NA (if not valid).
}
\description{
Compute a knowledge and understanding domain score based on the physical activity guideline, cardiorespiratory fitness means, muscular
strength and endurance means, sports skill and fill in the blanks scores. If one of the scores is missing or invalid, a weighted domain score will be
computed from the other four scores.
}
\details{
This function calls \code{\link[=validate_number]{validate_number()}}, \code{\link[=validate_integer]{validate_integer()}} and \code{\link[=validate_protocol_score]{validate_protocol_score()}} to ensure that the computed score is valid.
}
\examples{
get_ku_score(
  pa_guideline_score = c(1, 0, 1, 1, NA),
  cardiorespiratory_fitness_means_score = c(0, 1, "", 2, 1),
  muscular_strength_means_score = c(1, 1, 1, 0, 0),
  sports_skill_score = c(0, 0, 1, 0, 1),
  fill_in_the_blanks_score = c(5, 6, 3, 1, 2)
)

# [1] 7.000000 8.000000 6.666667 2.222222 4.444444

}
