% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_competence.R
\name{get_pacer_20m_laps}
\alias{get_pacer_20m_laps}
\title{Convert PACER shuttle run laps to their equivalent in 20-metre laps.}
\usage{
get_pacer_20m_laps(lap_distance = NA, laps_run = NA)
}
\arguments{
\item{lap_distance}{a numeric element or vector (valid values are 15 or 20).}

\item{laps_run}{a numeric element (if lap_distance = 15, valid values are between 1 and 298).}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Convert PACER shuttle run laps to their equivalent in 20-metre laps.
}
\examples{
get_pacer_20m_laps(
  lap_distance = c(15, 20, NA, "15"), 
  laps_run = rep(100, 4)
)
# [1]  77 100  NA  77

}
