% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_competence.R
\name{get_pacer_score}
\alias{get_pacer_score}
\title{Compute a PACER score based on the number of PACER laps run at a 20-metre distance.}
\usage{
get_pacer_score(pacer_laps_20m = NA)
}
\arguments{
\item{pacer_laps_20m}{a numeric element or vector.}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Compute a PACER score based on the number of PACER laps run at a 20-metre distance.
}
\examples{
get_pacer_score(c(1, 6, 12, 18, NA, 46, 31))
# [1]  0  1  2  3 NA  9  6

}
