% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_behaviour.R
\name{get_self_report_pa_score}
\alias{get_self_report_pa_score}
\title{Compute a score for a response to the self-reported physical activity question.}
\usage{
get_self_report_pa_score(x = NA)
}
\arguments{
\item{x}{a numeric (integer) element or vector representing the self-reported physical activity question (valid values are integers between 0 and 7).}
}
\value{
returns a numeric (integer) element (if valid) or NA (if not valid).
}
\description{
This function computes a score for a response to "During the past week (7 days), on how many days were you physically active for a total of at least 60
minutes per day? (all the time you spent in activities that increased your heart rate and made you breathe hard)?". This score is used to compute the
Daily Behaviour domain score.
}
\examples{
get_self_report_pa_score(c(1, 8, 3, 4, 5, 2, 7))

# [1]  0 NA  2  3  4  1  5

}
