% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_age}
\alias{validate_age}
\title{Check whether an element is numeric and between 8 and 12.}
\usage{
validate_age(x)
}
\arguments{
\item{x}{an element or vector.}
}
\value{
returns a numeric element with a value between 8 and 12 (if valid) or NA (if not valid).
}
\description{
Check whether an element is numeric and between 8 and 12.
}
\examples{
validate_age(c(7:13, "", NA, "12"))

# [1] NA  8  9 10 11 12 NA NA NA 12

}
