% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_character}
\alias{validate_character}
\title{Check whether a vector is a character and not of length zero or "".}
\usage{
validate_character(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Returns a character vector (if valid) or NA (if not valid).
}
\description{
This function checks whether a vector is a character and not of length zero or "".
}
\examples{
validate_character(c("beginning", "progressing", "achieving", "excelling", "", NA, 7))

# [1] "beginning"   "progressing" "achieving"   "excelling"   NA            NA 
# [7] "7"

}
