% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_domain_score}
\alias{validate_domain_score}
\title{Check whether a CAPL domain score is numeric and within a valid range.}
\usage{
validate_domain_score(x = NA, domain = NA)
}
\arguments{
\item{x}{an element or vector representing a CAPL domain score.}

\item{domain}{a character element representing domains within CAPL (valid values are "pc", "db", "mc", "ku"; valid values are not
case-sensitive).}
}
\value{
returns a numeric element (if valid) or NA (if not valid).
}
\description{
Check whether a CAPL domain score is numeric and within a valid range.
}
\examples{
validate_domain_score(
  x = c(34, 15, 10, 12.5, 25),
  domain = "pc"
)

# [1]   NA 15.0 10.0 12.5 25.0

}
