% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_gender}
\alias{validate_gender}
\title{Check whether an element can be classified as girl or boy.}
\usage{
validate_gender(x)
}
\arguments{
\item{x}{an element or vector.}
}
\value{
returns a character element with a value of "girl" or "boy" (if valid) or NA (if not valid).
}
\description{
Check whether an element can be classified as girl or boy.
}
\examples{
validate_gender(c("Girl", "GIRL", "g", "G", "Female", "f", "F", "", NA))
# [1] "girl" "girl" "girl" "girl" "girl" "girl" "girl" NA     NA

validate_gender(c("Boy", "BOY", "b", "B", "Male", "m", "M", "", NA))
# [1] "boy" "boy" "boy" "boy" "boy" "boy" "boy" NA    NA

}
