% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_integer}
\alias{validate_integer}
\title{Check whether a vector is an integer.}
\usage{
validate_integer(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Returns a numeric (integer) vector (if valid) or NA (if not valid).
}
\description{
This function checks whether a vector is an integer.
}
\examples{
validate_integer(c(2, 6, 3.3, "", NA, "6", "hello, world"))

# [1]  2  6 NA NA NA  6 NA

}
