% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_behaviour.R
\name{validate_steps}
\alias{validate_steps}
\title{Check whether daily steps as measured by a pedometer are valid.}
\usage{
validate_steps(steps = NA, wear_time = NA)
}
\arguments{
\item{steps}{a numeric (integer) element or vector representing the steps taken on a given day (valid values are between 1000 and 30000).}

\item{wear_time}{a numeric element representing the duration of time (in decimal hours) that a pedometer was worn on a given day (valid values are >= 10.0
hours).}
}
\value{
returns "yes" (if valid) or "no" (if not valid).
}
\description{
Check whether daily steps as measured by a pedometer are valid.
}
\examples{
validate_steps(
  steps = c(5400, 11001, 999, 31000, 8796),
  wear_time = c(10.1, 12.6, 10.2, 10.9, 9.5)
)

# [1] "yes" "yes" "no"  "no"  "no" 

}
